% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csgfix.r
\name{bank}
\alias{bank}
\title{changes data to Movebank format for use with ctmm package}
\usage{
bank(
  data,
  projectionType = "na",
  timeformat = "\%m/\%d/\%Y \%T",
  ct = TRUE,
  coordFix = FALSE
)
}
\arguments{
\item{data:}{dataframe, the input data}

\item{projectionType:}{boolean or character, FALSE if you don't want to convert lat/long to X/Y (if you want to use an existing X and Y). Accepted values: "North_America_Equidistant_Conic"}

\item{timeformat:}{character (format for use in as.POSIXct), the format Date_Time is in}

\item{ct:}{boolean, FALSE if you want time to be saved as a string in move bank format, TRUE if you want it as a POSIXct}

\item{coordFix:}{function, a function that converts lat and long at datetime to NAD83, inputs are vectors lat (numeric), long (numeric), and datetime (POSIXct). returns a data frame with columns "lat" and "long". FALSE if you don't want to change the coords at all.}
}
\description{
changes data to Movebank format for use with ctmm package
}
