% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csgfix.r
\name{generateRandomFlight}
\alias{generateRandomFlight}
\title{generates a fake flight data. it's all random so it doesn't behave like an actual bird.}
\usage{
generateRandomFlight(
  origin = c(0, 0),
  names = "csg",
  obs = 1000,
  maxturn = 15,
  timestep = 3600,
  timestart = Sys.time(),
  multi = 1/500,
  fakemisc = TRUE,
  n = FALSE
)
}
\arguments{
\item{origin:}{numeric vector of length 2, the coordinates where bird will start}

\item{names:}{character, "csg" or "movebank", depending on how you want the columns of the resulting data to be named}

\item{obs:}{numeric, number of observations to create}

\item{maxturn:}{numeric, the maximum change in heading in each step (measured in degrees). can turn maxturn left or right.}

\item{timestep:}{numeric, the number of seconds between each observation}

\item{timestart:}{POSIXct, the time you want the first observation to be}

\item{multi:}{numeric, the distance the bird travels in each observation (in change in lat/long) will be a random number in [0,multi)}

\item{fakemisc:}{boolean, if you want to fake the other columns needed for bank()}

\item{n:}{numeric, the Animal_ID or FALSE to randomize it}
}
\description{
generates a fake flight data. it's all random so it doesn't behave like an actual bird.
}
