% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csgfix.r
\name{pad}
\alias{pad}
\title{pads out an integer x to width with zeroes (to the left)}
\usage{
pad(x, width = 3)
}
\arguments{
\item{x:}{numeric or character, thing you want to pad}

\item{width:}{numberic, the width to pad to}
}
\description{
pads out an integer x to width with zeroes (to the left)
}
