% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csgfix.r
\name{projectionCylindricalEqualArea}
\alias{projectionCylindricalEqualArea}
\title{projects lat and long coordinates to x and y with cylindrical equal area projection}
\usage{
projectionCylindricalEqualArea(
  lat,
  long,
  type = NULL,
  standardlat = 0,
  standardlong = 0
)
}
\arguments{
\item{standardlong, }{standardlat: numeric, parameters for the projection}

\item{lat:}{numeric vector, latitude (lat and long should be the same length)}

\item{long:}{numeric vector, longitude}

\item{type:}{character, NULL to use your own parameters, or the name of a preset ("lambert","behrmann","gall",...)}
}
\description{
projects lat and long coordinates to x and y with cylindrical equal area projection
}
