% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csgfix.r
\name{projectionEquidistantConic}
\alias{projectionEquidistantConic}
\title{projects lat and long coordinates to x and y with equidistant conic (simple conic) projection}
\usage{
projectionEquidistantConic(
  lat,
  long,
  type = NULL,
  reflat = 0,
  reflong = 0,
  sp1 = 20,
  sp2 = 60
)
}
\arguments{
\item{reflat, }{reflong, sp1, sp2: numeric, parameters for the projection (reflat and reflong are reference latitude and longitude (origin), sp1 and sp2 are standard parallels)}

\item{lat:}{numeric vector, latitude (lat and long should be the same length)}

\item{long:}{numeric vector, longitude}

\item{type:}{character, NULL to use your own parameters, or the name of a preset}
}
\description{
projects lat and long coordinates to x and y with equidistant conic (simple conic) projection
}
