% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csgfix.r
\name{projectionNA}
\alias{projectionNA}
\title{projects lat and long coordinates to x and y with North American equidistant conic projection}
\usage{
projectionNA(lat, long, v2 = NULL)
}
\arguments{
\item{lat:}{numeric vector, latitude (lat and long should be the same length)}

\item{long:}{numeric vector, longitude}
}
\description{
projects lat and long coordinates to x and y with North American equidistant conic projection
}
