% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csgfix.r
\name{quickMove}
\alias{quickMove}
\title{gets a Move object from data}
\usage{
quickMove(
  data,
  proj = CRS("+init=epsg:4326 +proj=longlat +ellps=WGS84 +datum=WGS84"),
  removeDuplicatedTimestamps = FALSE
)
}
\arguments{
\item{data:}{dataframe, the input data for bank(data)}

\item{proj:}{CRS object, sent to move::move}

\item{removeDuplicatedTimestamps:}{boolean, sent to move::move}
}
\description{
gets a Move object from data
}
