% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csgfix.r
\name{quickTelemetry}
\alias{quickTelemetry}
\title{gets a Telemetry object from data, shortcut for ctmm::as.telemetry(quickMove(data, ...))}
\usage{
quickTelemetry(
  data,
  proj,
  timeformat,
  removeDuplicatedTimestamps,
  projectionType
)
}
\arguments{
\item{data:}{dataframe, the input data for bank(data)}

\item{proj:}{CRS object, sent to move::move}

\item{timeformat:}{character (format for use in as.POSIXct), the format Date_Time is in}

\item{removeDuplicatedTimestamps:}{boolean, sent to move::move}

\item{projectionType:}{character or function, sent to bank}
}
\description{
gets a Telemetry object from data, shortcut for ctmm::as.telemetry(quickMove(data, ...))
}
