% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csgfix.r
\name{stupidRnorm}
\alias{stupidRnorm}
\title{dumb thing to get random values with normal distribution, mean=0.5, sd=0.125, limited to [0,1]}
\usage{
stupidRnorm(n, mean = 0.5, sd = 0.125)
}
\arguments{
\item{n:}{numeric, number of numbers to generate}
}
\description{
dumb thing to get random values with normal distribution, mean=0.5, sd=0.125, limited to [0,1]
}
