% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csgfix.r
\name{bank}
\alias{bank}
\title{changes data to Movebank format for use with ctmm package}
\usage{
bank(
  data,
  projectionType = "na",
  timeformat = "\%m/\%d/\%Y \%T",
  ct = TRUE,
  tonad83 = FALSE
)
}
\arguments{
\item{data:}{dataframe, the input data}

\item{projectionType:}{character or function, NULL if you don't want to convert lat/long to X/Y (if you want to use an existing X and Y). Otherwise a character (the name of a projection), or a function(lat, long, datetime) where lat and long are vectors of numbers and datetime is a vector of POSIXct corresponding to the time of those observations. should return a data frame with two numeric columns, x and y}

\item{timeformat:}{character (format for use in as.POSIXct), the format Date_Time is in}

\item{ct:}{boolean, FALSE if you want time to be saved as a string in move bank format, TRUE if you want it as a POSIXct}

\item{tonad83:}{boolean, TRUE if you want to convert from WGS84 to NAD83 before projection (only used when projectionType is a character)}
}
\description{
changes data to Movebank format for use with ctmm package
}
