% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csgfix.r
\name{quickMove}
\alias{quickMove}
\title{gets a Move object from data}
\usage{
quickMove(
  data,
  proj = CRS("+ellps=WGS84 +datum=WGS84 +no_defs +proj=latlong"),
  timeformat = "\%m/\%d/\%Y \%H:\%M",
  removeDuplicatedTimestamps = FALSE,
  projectionType = NULL
)
}
\arguments{
\item{data:}{dataframe, the input data for bank(data)}

\item{proj:}{CRS object, sent to move::move (be sure to change this if you change projectionType)}

\item{timeformat:}{character (format for use in as.POSIXct), the format Date_Time is in}

\item{removeDuplicatedTimestamps:}{boolean, sent to move::move}

\item{projectionType:}{character or function, sent to bank}
}
\description{
gets a Move object from data
}
