% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csgfix.r
\name{quickMove}
\alias{quickMove}
\title{gets a Move object from data}
\usage{
quickMove(
  data,
 
    proj = CRS("+init=epsg:4326 +ellps=WGS84 +datum=WGS84 +no_defs +proj=eqdc +lat_1=20 +lat_2=60 +lat_0=0 +lon_0=0"),
  removeDuplicatedTimestamps = FALSE
)
}
\arguments{
\item{data:}{dataframe, the input data for bank(data)}

\item{proj:}{CRS object, sent to move::move}

\item{removeDuplicatedTimestamps:}{boolean, sent to move::move}
}
\description{
gets a Move object from data
}
